requirements.txt
django-extensions
mysqlclient
faker
numpy
autopep8